+++ src/main/java/com/brokerage/api/entity/Asset.java
@@ -15,6 +15,7 @@ import org.springframework.data.annotation.CreatedDate;
 import org.springframework.data.annotation.LastModifiedDate;
 import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.UUID;
 
@@ -38,6 +39,6 @@ public class Asset {
     @JoinColumn(name = "customer_id")
     private Customer customer;
     private String name;
-    private int size;
-    private int usableSize;
+    private BigDecimal size;
+    private BigDecimal usableSize;
 }
